refactor(lotto): 랜덤 클래스 분리
